Put everything back the way it was
I do not think that the CPU usage is due to anything we have in the game. I deselected just about everything and the usage sky rocketed the moment the bottom right of the progress bar said, "5/11 clustering" which has been there ever since we downloaded that cartoon cat from the asset store. (since then we have erased all files regarding that cat from our assets folder and I have no idea how it would still be effecting us. Currently my hope for the presentation tomorrow is that we let people play on the soft ware which work fine. (minus the occasional glicth that I spoken about with Whitworth)